add some storeowner rights

+++ laundry/src/main/java/com/example/laundry/models/order/Order.java
@@ -25,7 +25,7 @@ public class Order {
     private Long id;
 
     @Column
-    private BigDecimal totalAmount;
+    private Double totalAmount;
 
     @Column
     @Enumerated(EnumType.STRING)
